+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Unreleased
 
 *v0.0.1*
 
+**8/14/23**
+
+- [Add UI Component to ADD Block to Blockchain](#20)
+
 **8/1/23**
 
 - [Create Simple UI to Check API Readiness](#18)